research-agent/requirements.txt
anthropic==0.39.0
beautifulsoup4==4.12.3
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
httpx==0.27.2
jinja2==3.1.4
linkedin-scraper==2.11.1
lxml==5.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
rich==13.9.2
typer==0.21.1
beautifulsoup4